requirements.txt
Flask~=1.1
requests~=2.22
tensorflow~=2.2.0
pyyaml~=5.3.1
h5py~=2.10.0
Pillow~=7.1.2
image~=1.5.31
